Rename frontend to view
