+++ src/blocks/interactive-child/frontend.js
@@ -1,26 +1,8 @@
 import CounterContext from '../../context/counter';
 import ThemeContext from '../../context/theme';
-import { useContext } from '../../gutenberg-packages/wordpress-element';
+import { registerBlockType } from '../../gutenberg-packages/frontend';
+import View from './view';
 
-const Frontend = ({ blockProps, context }) => {
-	const theme = useContext(ThemeContext);
-	const counter = useContext(CounterContext);
-
-	return (
-		<div {...blockProps}>
-			<p>
-				Block Context from interactive parent - "bhe/interactive-title":{' '}
-				{context['bhe/interactive-title']}
-			</p>
-			<p>
-				Block Context from non-interactive parent -
-				"bhe/non-interactive-title":{' '}
-				{context['bhe/non-interactive-title']}
-			</p>
-			<p>React Context - "counter": {counter}</p>
-			<p>React Context - "theme": {theme}</p>
-		</div>
-	);
-};
-
-export default Frontend;
+registerBlockType('bhe/interactive-child', View, {
+	usesContext: [ThemeContext, CounterContext],
+});
