+++ src/blocks/non-interactive-parent/index.js
@@ -1,9 +1,12 @@
 import { registerBlockType } from '../../gutenberg-packages/wordpress-blocks';
 import metadata from './block.json';
-import edit from './edit';
-import frontend from './frontend';
+import Edit from './edit';
+import View from './view';
 import './style.scss';
 
 const { name } = metadata;
 
-registerBlockType(name, { edit, frontend });
+registerBlockType(name, {
+	edit: Edit,
+	view: View, // The Save component is derived from the View component.
+});
