+++ src/blocks/non-interactive-parent/view.js
@@ -1,8 +1,8 @@
-const Frontend = ({ attributes, blockProps, children }) => (
+const View = ({ attributes, blockProps, children }) => (
 	<div {...blockProps}>
 		<p className="title">{attributes.title}</p>
 		{children}
 	</div>
 );
 
-export default Frontend;
+export default View;
